configured terra monitors to run the service to monitor either v1 or v2 basset contacrts by setting env variable BASSET_CONTRACTS_VERSION #LSC-33
