+++ Makefile
@@ -5,7 +5,13 @@ dev_server: gen_client
 	go run main.go
 
 start:
-	docker-compose up -d --build
+	docker-compose --env-file ./docker/env/.lido_terra.env up -d --build
+
+start_testnet:
+	docker-compose --env-file ./docker/env/.lido_terra.testnet.env up --build -d
+
+start_mainnet:
+	docker-compose --env-file ./docker/env/.lido_terra.prod.env -p terra_monitors_mainnet up --build -d
 
 start-no-build:
 	docker-compose up -d
@@ -13,6 +19,11 @@ start-no-build:
 stop:
 	docker-compose down --remove-orphans
 
+stop_testnet:
+	docker-compose down --remove-orphans
+
+stop_mainnet:
+	docker-compose -p terra_monitors_mainnet down --remove-orphans
 
 test:
 	go test ./...
