+++ docker-compose.yml
@@ -3,7 +3,7 @@ services:
   grafana:
     image: grafana/grafana:8.0.5
     ports:
-      - "3000:3000"
+      - "${GRAFANA_PORT:-3000}:3000"
     depends_on:
       - prometheus
     env_file:
@@ -16,14 +16,10 @@ services:
       - grafana-storage:/var/lib/grafana
   renderer:
     image: grafana/grafana-image-renderer:latest
-    ports:
-      - 8081
     environment:
       ENABLE_METRICS: 'true'
   prometheus:
     image: prom/prometheus
-    ports:
-      - 9090:9090
     command:
       - --config.file=/etc/prometheus/prometheus.yml
     volumes:
@@ -33,28 +29,38 @@ services:
       - lido_terra
   lido_terra:
     build: .
-    env_file:
-      ./docker/env/.lido_terra.env
+    environment:
+      - LCD_ENDPOINT
+      - LCD_SCHEMES
+      - UPDATE_DATA_INTERVAL
+      - ADDRESSES_HUB_CONTRACT
+      - ADDRESSES_REWARD_CONTRACT
+      - ADDRESSES_BLUNA_TOKEN_INFO_CONTRACT
+      - ADDRESSES_VALIDATORS_REGISTRY_CONTRACT
+      - ADDRESSES_AIR_DROP_REGISTRY_CONTRACT
+      - ADDRESSES_UPDATE_GLOBAL_INDEX_BOT_ADDRESS
+      - ADDRESSES_REWARDS_DISPATCHER_CONTRACT
+      - BASSET_CONTRACTS_VERSION
     logging:
       driver: "json-file"
       options:
-        tag: "lido_terra"
+        tag: ${SERVICE_NAME:-lido_terra}
   loki:
     image: grafana/loki:2.3.0
     volumes:
       - loki-data:/loki
-    ports:
-      - "3100:3100"
     command: -config.file=/etc/loki/local-config.yaml
 
 
   promtail:
     image: grafana/promtail:2.3.0
+    environment:
+      - SERVICE_NAME=${SERVICE_NAME:-lido_terra}
     volumes:
       - /var/lib/docker/containers:/var/lib/docker/containers:ro
       - promtail-data:/var/lib/promtail/positions
       - ./docker/config/promtail/promtail.yml:/etc/promtail/promtail.yml
-    command: -config.file=/etc/promtail/promtail.yml
+    command: -config.file=/etc/promtail/promtail.yml --config.expand-env=true
 
 
 volumes:
